requirements.txt
flask
requests
pymongo
python-dotenv
groq
PyYAML
openai==0.28.0
langchain==0.1.0
langchain-community==0.0.12
langchain-core==0.1.10
qdrant-client
tiktoken==0.5.2
gunicorn
markupsafe
fpdf